+++ offchain/crates/solana-cli/src/command/shreds/payments.rs
@@ -212,6 +212,7 @@ impl PaymentsCommand {
                             | ShredSubscriptionInstructionData::InitializeClientSeat { .. }
                             | ShredSubscriptionInstructionData::RequestInstantSeatAllocation
                             | ShredSubscriptionInstructionData::RequestInstantSeatWithdrawal
+                            | ShredSubscriptionInstructionData::RequestProratedInstantSeatWithdrawal
                             | ShredSubscriptionInstructionData::SetValidatorClientRewardsProportion(
                                 _,
                             )